requirements.txt
fastapi
google-genai
# httpx
langdetect
psycopg2-binary
pypdf
# pytest
# pytest-asyncio
python-docx
sqlalchemy
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers==4.49.0
uvicorn
